+++ src/Client/Extensions/AuthorizationHeaderExtensions.cs
@@ -1,14 +1,15 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
+using System.Net.Http;
 using System.Net.Http.Headers;
 
-namespace System.Net.Http
+namespace IdentityModel.Client
 {
     /// <summary>
-    /// Extensions for HttpClient
+    /// Extensions for HttpRequestMessage
     /// </summary>
-    public static class HttpClientExtensions
+    public static class AuthorizationHeaderExtensions
     {
         /// <summary>
         /// Sets a basic authentication header.
@@ -20,7 +21,7 @@ namespace System.Net.Http
         {
             client.DefaultRequestHeaders.Authorization = new BasicAuthenticationHeaderValue(userName, password);
         }
-        
+
         /// <summary>
         /// Sets a basic authentication header for RFC6749 client authentication.
         /// </summary>
@@ -51,6 +52,49 @@ namespace System.Net.Http
         public static void SetBearerToken(this HttpClient client, string token)
         {
             client.SetToken("Bearer", token);
+        }
+
+        /// <summary>
+        /// Sets a basic authentication header.
+        /// </summary>
+        /// <param name="request">The HTTP request message.</param>
+        /// <param name="userName">Name of the user.</param>
+        /// <param name="password">The password.</param>
+        public static void SetBasicAuthentication(this HttpRequestMessage request, string userName, string password)
+        {
+            request.Headers.Authorization = new BasicAuthenticationHeaderValue(userName, password);
+        }
+        
+        /// <summary>
+        /// Sets a basic authentication header for RFC6749 client authentication.
+        /// </summary>
+        /// <param name="request">The HTTP request message.</param>
+        /// <param name="userName">Name of the user.</param>
+        /// <param name="password">The password.</param>
+        public static void SetBasicAuthenticationOAuth(this HttpRequestMessage request, string userName, string password)
+        {
+            request.Headers.Authorization = new BasicAuthenticationOAuthHeaderValue(userName, password);
+        }
+
+        /// <summary>
+        /// Sets an authorization header with a given scheme and value.
+        /// </summary>
+        /// <param name="request">The HTTP request message.</param>
+        /// <param name="scheme">The scheme.</param>
+        /// <param name="token">The token.</param>
+        public static void SetToken(this HttpRequestMessage request, string scheme, string token)
+        {
+            request.Headers.Authorization = new AuthenticationHeaderValue(scheme, token);
+        }
+
+        /// <summary>
+        /// Sets an authorization header with a bearer token.
+        /// </summary>
+        /// <param name="request">The HTTP request message.</param>
+        /// <param name="token">The token.</param>
+        public static void SetBearerToken(this HttpRequestMessage request, string token)
+        {
+            request.SetToken("Bearer", token);
         }
     }
 }
