+++ src/Client/Extensions/HttpClientDeviceFlowExtensions.cs
@@ -2,43 +2,38 @@
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
 using IdentityModel.Internal;
-using Newtonsoft.Json;
 using System;
-using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
-using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
 
 namespace IdentityModel.Client
 {
     /// <summary>
-    /// HttpClient extensions for dynamic registration
+    /// HttpClient extensions for OIDC userinfo
     /// </summary>
-    public static class HttpClientDynamicRegistrationExtensions
+    public static class HttpClientDeviceFlowExtensions
     {
         /// <summary>
-        /// Send a dynamic registration request.
+        /// Sends a userinfo request.
         /// </summary>
         /// <param name="client">The client.</param>
         /// <param name="request">The request.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <returns></returns>
-        public static async Task<RegistrationResponse> RegisterClientAsync(this HttpMessageInvoker client, DynamicClientRegistrationRequest request, CancellationToken cancellationToken = default)
+        public static async Task<DeviceAuthorizationResponse> RequestDeviceAuthorizationAsync(this HttpMessageInvoker client, DeviceAuthorizationRequest request, CancellationToken cancellationToken = default)
         {
+            request.Parameters.AddRequired(OidcConstants.AuthorizeRequest.ClientId, request.ClientId);
+            request.Parameters.AddOptional(OidcConstants.AuthorizeRequest.Scope, request.Scope);
+
             var httpRequest = new HttpRequestMessage(HttpMethod.Post, request.Address)
             {
-                Content = new StringContent(JsonConvert.SerializeObject(request.RegistrationRequest), Encoding.UTF8, "application/json")
+                Content = new FormUrlEncodedContent(request.Parameters)
             };
 
             httpRequest.Headers.Accept.Clear();
             httpRequest.Headers.Accept.Add(new MediaTypeWithQualityHeaderValue("application/json"));
-
-            if (request.Token.IsPresent())
-            {
-                httpRequest.SetBearerToken(request.Token);
-            }
 
             HttpResponseMessage response;
             try
@@ -47,10 +42,10 @@ namespace IdentityModel.Client
             }
             catch (Exception ex)
             {
-                return Response.FromException<RegistrationResponse>(ex);
+                return ProtocolResponse.FromException<DeviceAuthorizationResponse>(ex);
             }
 
-            return await Response.FromHttpResponseAsync<RegistrationResponse>(response);
+            return await ProtocolResponse.FromHttpResponseAsync<DeviceAuthorizationResponse>(response);
         }
     }
 }
