+++ src/Client/Messages/DeviceAuthorizationResponse.cs
@@ -1,14 +1,13 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
 namespace IdentityModel.Client
 {
     /// <summary>
     /// Models an OAuth device authorization response
     /// </summary>
-    /// <seealso cref="IdentityModel.Client.Response" />
-    public class DeviceAuthorizationResponse : Response
+    /// <seealso cref="IdentityModel.Client.ProtocolResponse" />
+    public class DeviceAuthorizationResponse : ProtocolResponse
     {
         /// <summary>
         /// Gets the device verification code.
