+++ src/Client/Messages/RegistrationResponse.cs
@@ -1,7 +1,6 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-
 #pragma warning disable 1591
 
 namespace IdentityModel.Client
@@ -9,8 +8,8 @@ namespace IdentityModel.Client
     /// <summary>
     /// Models an OpenID Connect dynamic client registration response
     /// </summary>
-    /// <seealso cref="IdentityModel.Client.Response" />
-    public class RegistrationResponse : Response
+    /// <seealso cref="IdentityModel.Client.ProtocolResponse" />
+    public class RegistrationResponse : ProtocolResponse
     {
         public string ErrorDescription         => Json.TryGetString("error_description");
         public string ClientId                 => Json.TryGetString(OidcConstants.RegistrationResponse.ClientId);
