+++ src/Client/Messages/TokenResponse.cs
@@ -1,16 +1,13 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using System;
-using System.Net;
-
 namespace IdentityModel.Client
 {
     /// <summary>
     /// Models a response from an OpenID Connect/OAuth 2 token endpoint
     /// </summary>
-    /// <seealso cref="IdentityModel.Client.Response" />
-    public class TokenResponse : Response
+    /// <seealso cref="IdentityModel.Client.ProtocolResponse" />
+    public class TokenResponse : ProtocolResponse
     {
         /// <summary>
         /// Gets the access token.
