+++ src/Client/Messages/TokenRevocationResponse.cs
@@ -1,15 +1,12 @@
 ﻿// Copyright (c) Brock Allen & Dominick Baier. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See LICENSE in the project root for license information.
 
-using System;
-using System.Net;
-
 namespace IdentityModel.Client
 {
     /// <summary>
     /// Models an OAuth 2.0 token revocation response
     /// </summary>
-    /// <seealso cref="IdentityModel.Client.Response" />
-    public class TokenRevocationResponse : Response
+    /// <seealso cref="IdentityModel.Client.ProtocolResponse" />
+    public class TokenRevocationResponse : ProtocolResponse
     { }
 }
